Refactor storage/blockdevice directory
