+++ storage/blockdevice/source/FlashSimBlockDevice.cpp
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-#include "FlashSimBlockDevice.h"
+#include "blockdevice/FlashSimBlockDevice.h"
 #include "platform/mbed_assert.h"
 #include "platform/mbed_atomic.h"
 #include <algorithm>
